exceptions and dto

+++ src/main/java/com/example/fms/controller/AccountController.java
@@ -1,10 +1,14 @@
 package com.example.fms.controller;
 
+import com.example.fms.dto.AccountDTO;
 import com.example.fms.entity.Account;
+import com.example.fms.entity.ResponseMessage;
 import com.example.fms.entity.User;
 import com.example.fms.service.AccountService;
 import com.example.fms.service.UserService;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.servlet.http.HttpServletRequest;
 import java.math.BigDecimal;
@@ -18,13 +22,19 @@ import java.util.Set;
 @RequestMapping("/account")
 public class AccountController {
     private final AccountService accountService;
+    private final UserService userService;
 
     AccountController(AccountService accountService, UserService userService) {
         this.accountService = accountService;
+        this.userService = userService;
     }
 
     @GetMapping("/get")
-    public List<Account> getAllByParam(HttpServletRequest request){
+    public List<Account> getAllByParam(HttpServletRequest request, Principal principal){
+        User user = userService.getByEmail(principal.getName());
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Authorize to get account information");
+        }
         String name = request.getParameter("name");
         String balanceLessThan = request.getParameter("balanceLessThan");
         String balanceGreaterThan = request.getParameter("balanceGreaterThan");
@@ -47,23 +57,54 @@ public class AccountController {
     }
 
     @GetMapping("/{accountId}")
-    public Account getAccount(@PathVariable("accountId") Long accountId) throws Exception{
-        return accountService.getAccountById(accountId);
+    public Account getAccount(@PathVariable("accountId") Long accountId, Principal principal) {
+        User user = userService.getByEmail(principal.getName());
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Authorize to get account information");
+        }
+        Account account = accountService.getAccountById(accountId);
+        if (account != null) {
+            return account;
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Account id " + accountId + " not found!");
     }
 
     @PostMapping("/add")
-    public Account add(@RequestBody Account newAccount, Principal principal) {
-        return accountService.create(newAccount, principal.getName());
+    public ResponseMessage add(@RequestBody AccountDTO accountDTO, Principal principal) {
+        User user = userService.getByEmail(principal.getName());
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Authorize to add account");
+        }
+        accountService.create(accountDTO, principal.getName());
+        return new ResponseMessage(HttpStatus.OK.value(), "Account successfully saved");
     }
 
-    @PutMapping("/update")
-    public Account updateAccount(@RequestBody Account newAccount, Principal principal) throws Exception{
-        return accountService.updateAccountById(newAccount, principal.getName());
+    @PutMapping("/update/{id}")
+    public ResponseMessage updateAccount(@RequestBody AccountDTO accountDTO, @PathVariable Long id, Principal principal){
+        User user = userService.getByEmail(principal.getName());
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Authorize to change account information");
+        }
+
+        Account account = accountService.updateAccountById(accountDTO, id, principal.getName());
+        if (account != null) {
+            return new ResponseMessage(HttpStatus.OK.value(), "Account successfully updated");
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Account id " + id + " not found!");
     }
 
-    @DeleteMapping("/delete/{id}")
-    public void deleteAccount(@PathVariable("id") Long id, Principal principal) {
-        accountService.deleteAccountById(id, principal.getName());
+    @DeleteMapping("/{id}")
+    public ResponseMessage deleteAccount(@PathVariable("id") Long id, Principal principal) {
+        User user = userService.getByEmail(principal.getName());
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Authorize to delete account information");
+        }
+
+        Account account = accountService.deleteAccountById(id, principal.getName());
+        if (account != null) {
+            return new ResponseMessage(HttpStatus.OK.value(), "Account successfully deleted");
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Account id " + id + " not found!");
     }
 
 }
