+++ src/main/java/com/example/fms/controller/DepartmentController.java
@@ -1,10 +1,14 @@
 package com.example.fms.controller;
 
-import com.example.fms.entity.Category;
+import com.example.fms.dto.DepartmentDTO;
 import com.example.fms.entity.Department;
+import com.example.fms.entity.ResponseMessage;
 import com.example.fms.entity.User;
 import com.example.fms.service.DepartmentService;
+import com.example.fms.service.UserService;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.servlet.http.HttpServletRequest;
 import java.security.Principal;
@@ -18,13 +22,19 @@ import java.util.Set;
 public class DepartmentController {
 
     private final DepartmentService departmentService;
+    private final UserService userService;
 
-    DepartmentController(DepartmentService service) {
+    DepartmentController(DepartmentService service, UserService userService) {
         this.departmentService = service;
+        this.userService = userService;
     }
 
     @GetMapping("/get")
-    public List<Department> getAllByParam(HttpServletRequest request) {
+    public List<Department> getAllByParam(HttpServletRequest request, Principal principal) {
+        User user = userService.getByEmail(principal.getName());
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Authorize to get department information");
+        }
         String name = request.getParameter("name");
         String dateAfter = request.getParameter("dateAfter");
         String dateBefore = request.getParameter("dateBefore");
@@ -42,23 +52,51 @@ public class DepartmentController {
     }
 
     @PostMapping("/add")
-    public Department addDepartment (@RequestBody Department newDepartment, Principal principal) {
-        return departmentService.addDepartment(newDepartment, principal.getName());
+    public Department addDepartment (@RequestBody DepartmentDTO departmentDTO, Principal principal) {
+        User user = userService.getByEmail(principal.getName());
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Authorize to add department");
+        }
+        return departmentService.addDepartment(departmentDTO, principal.getName());
     }
 
     @GetMapping("/{id}")
-    public Department getDepartment(@PathVariable Long id) throws Exception {
-        return departmentService.getDepartmentById(id);
+    public ResponseMessage getDepartment(@PathVariable Long id, Principal principal) {
+        User user = userService.getByEmail(principal.getName());
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Authorize to get department information");
+        }
+        Department department = departmentService.getDepartmentById(id);
+        if (department != null) {
+            return new ResponseMessage(HttpStatus.OK.value(), "Department successfully saved");
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Department id " + id + " not found!");
     }
 
-    @PutMapping ("/update")
-    public Department updateDepartment(@RequestBody Department newDepartment, Principal principal) throws Exception{
-        return departmentService.updateDepartmentById(newDepartment, principal.getName());
+    @PutMapping ("/update/{id}")
+    public ResponseMessage updateDepartment(@RequestBody DepartmentDTO departmentDTO, @PathVariable Long id, Principal principal){
+        User user = userService.getByEmail(principal.getName());
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Authorize to change department information");
+        }
+        Department department = departmentService.updateDepartmentById(departmentDTO, id, principal.getName());
+        if (department != null) {
+            return new ResponseMessage(HttpStatus.OK.value(), "Department successfully updated");
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Department id " + id + " not found!");
     }
 
     @DeleteMapping("/delete/{id}")
-    public boolean deleteDepartment(@PathVariable Long id, Principal principal) {
-        return departmentService.deleteDepartmentById(id, principal.getName());
+    public ResponseMessage deleteDepartment(@PathVariable Long id, Principal principal) {
+        User user = userService.getByEmail(principal.getName());
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Authorize to delete department information");
+        }
+        Department department = departmentService.deleteDepartmentById(id, principal.getName());
+        if (department != null) {
+            return new ResponseMessage(HttpStatus.OK.value(), "Department successfully deleted");
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Department id " + id + " not found!");
     }
 
 }
